More surface work, but in progress, not fully functional
There is some leg work to do in `propagate` in order to reach the point where we score on surfaces: we must check if we cross an interface, then move to it, then decide whetehr or not it is transmitted or reflected, score and leave if you have to.
It is important to separate two things: `contains()` and `intersect()`. The two functions serve different purposes because we may not want to compute surface stats all the time and also we don't want to rely on checking intersections to decide if a photon is inside or outside.