Refactoring
Really not sure where to put the reflexion and refraction methods: they depend on 1) surface, 2) photon, 3) material, and they only class that really knows about all these objects (and is the cause of all this) is Geometry.
They *feel* natural in `Photon`, but they are awkward to use because of  the plane of incidence normal and angle of incidence that need to be recalculated.

In Apple's library (SceneKit) , they have an object called `Intersect` that allows one to manipulate all of these parameters in a single object, even though it may be tempting to just stuff everything into `Surface`.