UnitVector was not properly normalized when initialized when non-unit vector
The override of `abs()` assumed the vector would be nearly nomralized, which is not always the case.
Also, normalizedCrossedProduct  really needs to return a value less than 1.0 so we really force the result to a UnitVector because when using it with angleWith, asin(phi) would blow up sometimes due to round off errors.