bug fix: nextExitInterface() was returning incorrect distance
When we are against a surface (i.e. against its normal), we will never cross it.  Picking the surface element at the end checked that when sitting exctaly on a surface but incorrectly set the distance to minDistance instead and then the photon could not move anymore.  This is fixed.