Update VectorOperators.c : Fix parameter types
There are two different calls in the code. First, rotate photon direction around its 'e_r' and vice versa. This first case requires two global pointers to float3 vectors as argument. This signature is in `rotateAroundAxisGlobal`. The next use case is to reflec or refract the photon direction around the Fresnel Incidence plane (locally created). This case requires a global float3 pointer as first argument and a local float3 pointer as second argument. This is done in `rotateAround`.