Attempt #1 at working around OpenCL bugs
The current bug on M2 with OpenCL is related to the smoothNormal kernel.  I added a check for an environment variable to avoid smoothing, but I also had to modify intersection.c and now some other tests don't pass.

The problem in intersection.c is related to passing __global pointers around but one function creates a __local INtersection, so passing the pointer makes no sense and the compiler complains.  We need to find a way to either pass Intersection directly( which I tried to do) or always refer to the __global intersections array.  I tried to make it work by passing Intersections around instead of passing the pointers, but not sure if this is functional (the tests don't pass but I dont have a proof they passed before).