Update testCompactStructures.py
I now have a class that creates a large array of contiguous data laid out as a dtype.struct (user-defined).  Then a CompactObject allows to access the struct as a an object that may have other functionality.

This is convenient to create large arrays for OpenCL for data properly structured for GPU upload while allowing easy management and creation of this data.